fix - changed LIC2's EPSILON to double to match other constants (#30)

+++ src/main/java/com/example/ConditionCheck.java
@@ -19,7 +19,7 @@ public class ConditionCheck {
         return new boolean[] {
             LIC0.LIC0(points, LENGTH1),
             LIC1.LIC1(points, RADIUS1),
-            LIC2.LIC2(points, (float) EPSILON),
+            LIC2.LIC2(points, EPSILON),
             LIC3.LIC3(points, AREA1),
             LIC4.LIC4(points, Q_PTS, QUADS),
             LIC5.LIC5(points),
